Refactor: Status Enum 이름 변경
심부름의 상태 (모집중, 진행중, 완료)를 표시하는 enum class의 이름을 Status에서 ErrandStatus로 변경.
이유는 현재 현황페이지 관련 클래스를 Status로 하기 위해 구분하기 위해서
