+++ server/appteam/src/main/java/com/pknuErrand/appteam/Enum/ErrandStatus.java
@@ -1,5 +1,5 @@
 package com.pknuErrand.appteam.Enum;
 
-public enum Status {
+public enum ErrandStatus {
     RECRUITING, IN_PROGRESS, DONE;
 }
