+++ server/appteam/src/main/java/com/pknuErrand/appteam/controller/status/StatusController.java
@@ -1,0 +1,8 @@
+package com.pknuErrand.appteam.controller.status;
+
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class StatusController {
+}
