+++ server/appteam/src/main/java/com/pknuErrand/appteam/dto/errand/defaultDto/ErrandRequestDto.java
@@ -1,6 +1,6 @@
 package com.pknuErrand.appteam.dto.errand.defaultDto;
 
-import com.pknuErrand.appteam.Enum.Status;
+import com.pknuErrand.appteam.Enum.ErrandStatus;
 import com.pknuErrand.appteam.domain.member.Member;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -33,7 +33,7 @@ public class ErrandRequestDto { // to Entity
 
     private Boolean isCash;
 
-    private Status status;
+    private ErrandStatus errandStatus;
 
     private Member erranderNo; // 심부름꾼
