+++ server/appteam/src/main/java/com/pknuErrand/appteam/dto/errand/getDto/ErrandPaginationRequestVo.java
@@ -1,9 +1,8 @@
 package com.pknuErrand.appteam.dto.errand.getDto;
 
-import com.pknuErrand.appteam.Enum.Status;
+import com.pknuErrand.appteam.Enum.ErrandStatus;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.Setter;
 
 @AllArgsConstructor
 @Getter
@@ -11,5 +10,5 @@ public class ErrandPaginationRequestVo {
     private Long pk;
     private String cursor;
     private int limit;
-    private Status status;
+    private ErrandStatus errandStatus;
 }
